refactoring conversions module layout
