+++ .readthedocs.yaml
@@ -17,6 +17,8 @@ sphinx:
 # Optionally declare the Python requirements required to build your docs
 python:
   install:
+    - requirements: requirements.txt
+    - requirements: requirements-dev.txt
     - method: pip
       path: .
       extra_requirements:
