+++ qbraid/transpiler/conversions/braket/braket_to_pytket.py
@@ -9,7 +9,7 @@
 # THERE IS NO WARRANTY for the qBraid-SDK, as per Section 15 of the GPL v3.
 
 """
-Module containing functions to convert between Amazon Braket and PyTKET.
+Module containing function to convert from Amazon Braket to PyTKET.
 
 """
 
@@ -36,20 +36,3 @@ def braket_to_pytket(circuit: "braket.circuits.Circuit") -> "pytket.circuit.Circ
     from pytket.extensions.braket import braket_convert  # type: ignore
 
     return braket_convert.braket_to_tk(circuit)
-
-
-@requires_extras("pytket.extensions.braket")
-def pytket_to_braket(circuit: "pytket.circuit.Circuit") -> "braket.circuits.Circuit":
-    """Returns an Amazon Braket circuit equivalent to the input pytket circuit.
-
-    Args:
-        circuit (pytket.circuit.Circuit): PyTKET circuit to convert to Braket circuit.
-
-    Returns:
-        braket.circuits.Circuit: Braket circuit equivalent to input pytket circuit.
-    """
-    # pylint: disable-next=import-outside-toplevel
-    from pytket.extensions.braket import braket_convert  # type: ignore
-
-    braket_circuit, _, _ = braket_convert.tk_to_braket(circuit)
-    return braket_circuit
