+++ qbraid/transpiler/conversions/cirq/cirq_to_braket.py
@@ -38,10 +38,10 @@ except ImportError:
 import qbraid.programs.libs.cirq
 from qbraid.transpiler.exceptions import CircuitConversionError
 
-from .custom_instr import C as BKControl
+from .braket_custom import C as BKControl
 
 
-def _cirq_to_braket(circuit: Circuit) -> BKCircuit:
+def cirq_to_braket(circuit: Circuit) -> BKCircuit:
     """Returns a Braket circuit equivalent to the input Cirq circuit.
 
     Args:
