+++ qbraid/transpiler/conversions/qasm3/qasm3_to_pyqir.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2023 qBraid
+# Copyright (C) 2024 qBraid
 #
 # This file is part of the qBraid-SDK
 #
@@ -7,23 +7,3 @@
 # See the LICENSE file in the project root or <https://www.gnu.org/licenses/gpl-3.0.html>.
 #
 # THERE IS NO WARRANTY for the qBraid-SDK, as per Section 15 of the GPL v3.
-
-"""
-Cirq conversions
-
-.. currentmodule:: qbraid.transpiler.conversions.cirq
-
-Functions
-----------
-
-.. autosummary::
-   :toctree: ../stubs/
-
-   cirq_to_braket
-   cirq_to_qasm2
-   cirq_to_pyquil
-
-"""
-from .cirq_to_braket import cirq_to_braket
-from .cirq_to_pyquil import cirq_to_pyquil
-from .cirq_to_qasm2 import cirq_to_qasm2
