+++ qbraid/transpiler/conversions/qiskit/__init__.py
@@ -19,10 +19,11 @@ Functions
 .. autosummary::
    :toctree: ../stubs/
 
-   qasm3_to_qiskit
-   qiskit_to_qasm3
-   qasm2_to_qiskit
+   qiskit_to_braket
    qiskit_to_qasm2
+   qiskit_to_qasm3
 
 """
-from .conversions_qasm import qasm2_to_qiskit, qasm3_to_qiskit, qiskit_to_qasm2, qiskit_to_qasm3
+from .qiskit_to_braket import qiskit_to_braket
+from .qiskit_to_qasm2 import qiskit_to_qasm2
+from .qiskit_to_qasm3 import qiskit_to_qasm3
