+++ tests/fixtures/cirq/gates.py
@@ -30,7 +30,8 @@ from cirq.ops.matrix_gates import MatrixGate
 from cirq.ops.swap_gates import ISwapPowGate, SwapPowGate
 from cirq.ops.three_qubit_gates import CCXPowGate, CCZPowGate
 
-from qbraid.transpiler.conversions.cirq.custom_ops import U3Gate, _give_cirq_gate_name
+from qbraid.transpiler.conversions.braket.braket_to_cirq import _give_cirq_gate_name
+from qbraid.transpiler.conversions.qasm2.cirq_custom import U3Gate
 
 cirq_gates = {
     "H": HPowGate,
