+++ tests/transpiler/braket/test_to_braket_measure.py
@@ -15,7 +15,7 @@ Unit tests for converting Cirq circuits with MeasurementGate(s) to Braket circui
 import cirq
 
 from qbraid.interface import circuits_allclose
-from qbraid.transpiler.conversions.braket.conversions_cirq import cirq_to_braket
+from qbraid.transpiler.conversions.cirq import cirq_to_braket
 
 
 def alpha_pump(sys, env):
