+++ tests/transpiler/pyquil/test_conversions_pyquil.py
@@ -22,7 +22,8 @@ from pyquil.gates import CNOT, CZ, RX, RZ, H, X, Y, Z
 from pyquil.noise import _decoherence_noise_model, _get_program_gates, apply_noise_model
 
 from qbraid.interface import circuits_allclose
-from qbraid.transpiler.conversions.pyquil.conversions_cirq import cirq_to_pyquil, pyquil_to_cirq
+from qbraid.transpiler.conversions.cirq import cirq_to_pyquil
+from qbraid.transpiler.conversions.pyquil import pyquil_to_cirq
 from qbraid.transpiler.exceptions import CircuitConversionError
