+++ tests/transpiler/pytket/test_conversions_pytket.py
@@ -19,7 +19,7 @@ from pytket.circuit import Circuit as TKCircuit
 from pytket.qasm import circuit_to_qasm_str
 
 from qbraid.interface import circuits_allclose
-from qbraid.transpiler.conversions.cirq import qasm2_to_cirq
+from qbraid.transpiler.conversions.qasm2 import qasm2_to_cirq
 from qbraid.transpiler.converter import transpile
 
 from ..cirq_utils import _equal
