+++ tests/transpiler/qasm/test_qasm2_to_qasm3.py
@@ -26,10 +26,7 @@ from qbraid.interface import circuits_allclose
 from qbraid.interface.random import random_circuit
 from qbraid.interface.random.qasm3_random import _qasm3_random
 from qbraid.programs import load_program
-from qbraid.transpiler.conversions.openqasm3.conversions_qasm import (
-    _get_qasm3_gate_defs,
-    qasm2_to_qasm3,
-)
+from qbraid.transpiler.conversions.qasm2.qasm2_to_qasm3 import _get_qasm3_gate_defs, qasm2_to_qasm3
 
 from ...fixtures.qasm3.circuits import qasm3_bell, qasm3_shared15
